feat: restructured init class + set init
